+++ euhedral-data-structures/src/test/java/euhedral/queues/PartitionedUnboundedSpmcArrayQueueTest.java
@@ -76,7 +76,7 @@ class PartitionedUnboundedSpmcArrayQueueTest {
             for (int i = 0; i < 8; i++) {
                 exec.submit(() -> {
                     while (drained.sum() < batch) {
-                        int count = q.drain(consumer, 4096);
+                        long count = q.drain(consumer, 4096);
                         drained.add(count);
                         Thread.yield();
                     }
