+++ euhedral-data-structures/src/test/java/euhedral/queues/PartitionedUnboundedSpscArrayQueueTest.java
@@ -73,7 +73,7 @@ class PartitionedUnboundedSpscArrayQueueTest {
 
             exec.submit(() -> {
                 while (drained.sum() < batch) {
-                    int count = q.drain(consumer, 4096);
+                    long count = q.drain(consumer, 4096);
                     drained.add(count);
                     Thread.yield();
                 }
